# openapi.yaml
openapi: 3.0.0
info:
  title: File Management API
  version: 1.0.0
paths:
  /writeFile:
    post:
      summary: Writes content to a file
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                  description: Path to the file
                fileContent:
                  type: string
                  description: Content to write to the file
      responses:
        "200":
          description: File written successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  serverId:
                    type: string
                  requestId:
                    type: string
                  data:
                    type: object
        "405":
          description: Method not allowed
        "500":
          description: Internal Server Error
  /readFile:
    get:
      summary: Reads content from a file
      parameters:
        - name: filePath
          in: query
          required: true
          description: Path to the file
          schema:
            type: string
      responses:
        "200":
          description: File read successfully
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: File not found
        "405":
          description: Method not allowed
        "500":
          description: Internal Server Error
  /listFiles:
    get:
      summary: Lists files in a directory
      parameters:
        - name: dirPath
          in: query
          required: true
          description: Path to the directory
          schema:
            type: string
      responses:
        "200":
          description: Files listed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "405":
          description: Method not allowed
        "500":
          description: Internal Server Error
  /deleteFile:
    delete:
      summary: Deletes a file
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                  description: Path to the file to delete
      responses:
        "200":
          description: File deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  serverId:
                    type: string
                  requestId:
                    type: string
                  data:
                    type: object
        "404":
          description: File not found
        "405":
          description: Method not allowed
        "500":
          description: Internal Server Error
  /generateFiles:
    post:
      summary: Generates multiple 10MB files in the specified directory
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                dirPath:
                  type: string
                  description: Directory path where the files should be generated
                sizeInMB:
                  type: integer
                  description: Total size in MB. Multiple 10MB files will be generated to achieve this.
      responses:
        "200":
          description: Files generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  serverId:
                    type: string
                  requestId:
                    type: string
                  data:
                    type: object
        "400":
          description: Bad Request (invalid input)
        "405":
          description: Method not allowed
        "500":
          description: Internal Server Error